requirements.txt
Django==4.2
djangorestframework==3.16.1
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.12.1
django-bootstrap-v5==1.0.11
psycopg2-binary==2.9.3
python-dotenv==1.0.0
requests==2.31.0
pandas==1.3.5
pytz==2021.3